Revert "Projects: add endpoints to manage mirrored repositories + packages"
